==> globalprep/spp/v2023/_setup/3_set_up_iucn_trends.Rmd <==
---
title: 'Set up IUCN extinction risk trends'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../../workflow/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---



```{r}
knitr::opts_chunk$set(echo = TRUE, eval=FALSE)
```


``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'Figs/',
                      echo = TRUE, message = FALSE, warning = FALSE, eval = FALSE)

library(dplyr)
library(rgdal)
library(raster)
library(here)
library(ggridges)
library(ggplot2)

source(here('workflow/R/common.R'))

goal     <- 'spp'
scenario <- 'v2022'
dir_goal <- here('globalprep', goal, scenario)

### goal specific folders and info
dir_data  <- file.path(dir_goal, '_data')
dir_setup <- file.path(dir_goal, '_setup')
dir_anx  <- file.path(dir_M, 'git-annex/globalprep')
dir_goal_anx <- file.path(dir_anx, goal, scenario, 'spp_risk_dists')

# ### provenance tracking
#library(provRmd); prov_setup()

source(file.path(dir_setup, 'common_fxns.R'))

```

# Summary

Determine IUCN extinction risk trends for marine species by regressing time series assessments against population trends.  This will also be explored for regionally assessed species.

# Data Sources

### IUCN Red List

# Methods

## Get IUCN historical assessments for IUCN marine species

This process is accomplished in `1_set_up_iucn_risk.Rmd` and `2_set_up_iucn_risk_regional.Rmd`.  Resulting timeseries of historical assessments are available in `data/iucn_risk_timeseries_2022-1.csv` and `data/iucn_risk_ts_rgn_2022-1.csv` (or with updated version numbers as applicable).

## Get all population trends for assessed marine species

Both global and regional trends are gleaned from the API.

``` {r get_trends_global}

trends_file <- file.path(dir_setup, 'int/trend_calcs',
                         sprintf('iucn_risk_ts_and_trends_%s.csv', api_version))

if(!file.exists(trends_file)) {
  
  spp_risk_file <- file.path(dir_data,
                             sprintf('iucn_risk_current_%s.csv', api_version))

  ### Filter to just the valid species and get narratives.  The cols_only bit
  ### is to prevent it from reading it as a double and failing because of scientific notation.
  spp_ids_valid <- read_csv(spp_risk_file, col_types = cols_only(iucn_sid = 'i')) %>%
    filter(!is.na(iucn_sid)) %>%
    .$iucn_sid %>%
    unique()
  
  ### /api/v3/species/narrative/id/:id?token='YOUR TOKEN'
  spp_narr_url <- 'http://apiv3.iucnredlist.org/api/v3/species/narrative/id/%s?token=%s'

 
  chunk_size <- 2000 ## pick a chunk size
  n_chunks <- ceiling(length(spp_ids_valid)/chunk_size)
  
  for(j in 1:n_chunks) { 
    
    options(timeout = 5000000)  
    ###j <- 10 : 10 not working
    spp_index <- c( ((j - 1) * chunk_size + 1) : min((j * chunk_size), length(spp_ids_valid)) )

    
    chunk_file <- file.path(dir_goal_anx, 'tmp', 
                    sprintf('spp_narr_chunk_%s_%s.csv', 
                            min(spp_index), max(spp_index)))

    if(!file.exists(chunk_file)) {
      cat_msg('Getting narrative info for species ', min(spp_index), ' to ', max(spp_index))
      
      spp_ids_chunk <- spp_ids_valid[spp_index]
      spp_narr_chunk <- mc_get_from_api(spp_narr_url, spp_ids_chunk, api_key, cores = 12, delay = .5)
      
      cat_msg('... found ', nrow(spp_narr_chunk), ' narrative rows for these species')
      
      write_csv(spp_narr_chunk, chunk_file)
       
    } else {
      
      cat_msg('Chunk file ', chunk_file, ' already exists; skipping these spp')
      
    }
  }  
  
  
  
spp_narr_chunk_files <- list.files(file.path(dir_goal_anx, 'tmp'), 
                                    pattern = 'spp_narr_chunk', 
                                    full.names = TRUE)
  
  spp_narr_df <- lapply(spp_narr_chunk_files, FUN = function(x) {
    read.csv(x)}) %>%
    bind_rows() %>%
    rename(iucn_sid = name, pop_trend = populationtrend) %>%
    dplyr::select(iucn_sid, pop_trend) %>%
    arrange(iucn_sid) 

  
    spp_trend_global <- spp_narr_df %>%
    bind_rows() %>%
    mutate(iucn_sid = as.integer(iucn_sid)) %>%
    mutate(iucn_version = api_version)

  #save this file for safe keeping...
    write_csv(spp_trend_global, file.path(dir_setup, 'int/trend_calcs',
                         sprintf('global_trends_%s.csv', api_version)))

  
  ### Repeat for regional
  
  spp_risk_file <- file.path(dir_data,
                             sprintf('iucn_risk_rgn_current_%s.csv', api_version))

  ### Filter to just the valid species and get narratives.  The cols_only bit
  ### is to prevent it from reading it as a double and failing because of scientific notation.
  spp_ids_valid <- read_csv(spp_risk_file, col_types = cols_only(iucn_sid = 'i', iucn_rgn = 'c')) %>%
    filter(!is.na(iucn_sid)) %>%
    distinct() %>% filter(!is.na(iucn_rgn))
  
  ### /api/v3/species/narrative/id/:id/region/:region_identifier?token='YOUR TOKEN'
  spp_narr_url <- 'http://apiv3.iucnredlist.org/api/v3/species/narrative/id/%s/region/RGN_ID?token=%s'
  
  spp_narr <- vector('list', length = nrow(spp_ids_valid))
  for(i in c(1:nrow(spp_ids_valid))) { # i <- 122
    spp_id <- spp_ids_valid$iucn_sid[i]
    rgn_id <- spp_ids_valid$iucn_rgn[i]
    cat_msg(i, 'of', nrow(spp_ids_valid), ': Trying spp', spp_id, 'from', rgn_id, '\n')
    spp_narr_rgn_url <- spp_narr_url %>% str_replace('RGN_ID', rgn_id)
    
    spp_narr_tmp <- get_from_api(spp_narr_rgn_url, spp_id, api_key, delay = .1)
    spp_narr[[i]] <- spp_narr_tmp
  }
  

  spp_trend_reclass <- spp_narr %>%
    lapply(FUN = function(x) dplyr::select(x, iucn_sid = name, 
                                    iucn_rgn = region_identifier, 
                                    pop_trend = result.populationtrend))
  ### have to select out the columns of interest b/c one of the other cols was
  ### switching between character and logical, so bind_rows wouldn't work
  
  spp_trend_rgn <- spp_trend_reclass %>%
    bind_rows() %>%
    mutate(iucn_sid = as.integer(iucn_sid)) %>%
    mutate(iucn_version = api_version)
  
    #save this file for safe keeping...
  write_csv(spp_trend_rgn, file.path(dir_setup, 'int/trend_calcs',
                         sprintf('rgn_trends_%s.csv', api_version)))
  
  
  
  spp_trend_all <- spp_trend_global %>%
    mutate(iucn_rgn = 'global') %>%
    bind_rows(spp_trend_rgn)
  
  write_csv(spp_trend_all, trends_file)

}

```


## Regress species with multiple recent historic assessments against population trends

From data gathered in the execution of SPP v2022, collect species data on past assessments (category) and population trend.  Combine and save.  Note that trend has to be assumed to be related to the most recent assessment only - not included in historical assessment API call.  We will need both a text description of trend (i.e. stable, increasing, or decreasing) and multiple assessments.

For this analysis we will focus only on assessments performed since 1991, the year of v1.0 Red List.

``` {r gather_data_clip_from_1991}

spp_risk_ts_file <- file.path(dir_data,
                              sprintf('iucn_risk_timeseries_%s.csv', api_version))
spp_risk_rgn_ts_file <- file.path(dir_data,
                              sprintf('iucn_risk_ts_rgn_%s.csv', api_version))

spp_risk_ts <- read_csv(spp_risk_ts_file) %>%
  mutate(iucn_rgn = 'global') %>%
  bind_rows(read_csv(spp_risk_rgn_ts_file) %>%
              rename(cat_ts = rgn_cat_ts, cat_ts_score = rgn_cat_ts_score))

cat_trend_91 <- read_csv(trends_file) %>%
  left_join(spp_risk_ts, by = c('iucn_sid', 'iucn_version', 'iucn_rgn')) %>%
  distinct() %>%
  filter(year >= 1991 & !is.na(cat_ts_score) & !is.na(pop_trend)) %>%
  group_by(iucn_sid, iucn_rgn) %>%
  mutate(n_assess = n()) %>%
  ungroup()

lm_91 <- cat_trend_91 %>%
  filter(n_assess >= 2) %>%
  group_by(iucn_sid, iucn_rgn) %>%
  do(calc_trend = lm( cat_ts_score ~ year, data = .)[['coefficients']][['year']]) %>%
  mutate(calc_trend = round(calc_trend, 5))

trend_91 <- cat_trend_91 %>%
  filter(n_assess > 1) %>%
  group_by(iucn_sid, iucn_rgn) %>%
  summarize(years  = paste(year, collapse = ', '),
            cat_ts = paste(cat_ts, collapse = ', '),
            scores = paste(cat_ts_score, collapse = ', '),
            pop_trend_desc = first(pop_trend)) %>%
  ungroup() %>%
  left_join(lm_91, by = c('iucn_sid', 'iucn_rgn'))

write_csv(trend_91, file.path(dir_setup, 'int/trend_calcs/trend_lm_vs_pop_91.csv'))

DT::datatable(trend_91)
```

``` {r gather_data_clip_from_2001}

cat_trend_01 <- read_csv(trends_file) %>%
  left_join(spp_risk_ts, by = c('iucn_sid', 'iucn_version', 'iucn_rgn')) %>%
  distinct() %>%
  filter(year >= 2001 & !is.na(cat_ts_score) & !is.na(pop_trend)) %>%
  group_by(iucn_sid, iucn_rgn) %>%
  mutate(n_assess = n()) %>%
  ungroup()

lm_01 <- cat_trend_01 %>%
  filter(n_assess >= 2) %>%
  group_by(iucn_sid, iucn_rgn) %>%
  do(calc_trend = lm( cat_ts_score ~ year, data = .)[['coefficients']][['year']]) %>%
  mutate(calc_trend = round(calc_trend, 5))

trend_01 <- cat_trend_01 %>%
  filter(n_assess > 1) %>%
  group_by(iucn_sid, iucn_rgn) %>%
  summarize(years  = paste(year,   collapse = ', '),
            cat_ts = paste(cat_ts, collapse = ', '),
            scores = paste(cat_ts_score, collapse = ', '),
            pop_trend_desc = first(pop_trend)) %>%
  ungroup() %>%
  left_join(lm_01, by = c('iucn_sid', 'iucn_rgn'))

write_csv(trend_01, file.path(dir_setup, 'int/trend_calcs/trend_lm_vs_pop_01.csv'))

DT::datatable(trend_01)

```

Note that the category scores are higher for higher-risk assessments; so a decreasing population should result in an increasing risk score and vice versa.

``` {r explore_rels}

trend_91 <- read_csv(file.path(dir_setup, 'int/trend_calcs/trend_lm_vs_pop_91.csv')) %>%
  filter(!is.na(pop_trend_desc)) %>%
  mutate(pop_trend = case_when(pop_trend_desc == 'decreasing' ~ -1,
                               pop_trend_desc == 'increasing' ~  1,
                               pop_trend_desc == 'stable'     ~  0,
                               TRUE                           ~ NA_real_))
  ### NOTE: here we're coding the trend in *population* not risk.

trend_regr_91 <- lm(calc_trend ~ pop_trend, data = trend_91)

trend_coeffs_91 <- summary(trend_regr_91)


### BASED ON FILTERING SINCE 1991 (IUCN Red List v1.0)
# Call:
# lm(formula = calc_trend ~ pop_trend, data = trend_91)
# 
# Residuals:
#      Min       1Q   Median       3Q      Max 
# -0.59919 -0.00436  0.00081  0.00524  0.19564 
# 
# Coefficients:
#               Estimate Std. Error t value Pr(>|t|)    
# (Intercept) -0.0008096  0.0006427  -1.260    0.208    
# pop_trend   -0.0051707  0.0007742  -6.679 3.32e-11 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.02122 on 1594 degrees of freedom
#   (1029 observations deleted due to missingness)
# Multiple R-squared:  0.02722,	Adjusted R-squared:  0.02661 
# F-statistic:  44.6 on 1 and 1594 DF,  p-value: 3.322e-11


trend_plot_91 <- ggplot(trend_91, aes(x = pop_trend, y = calc_trend, group = pop_trend)) +
  geom_violin(color = 'grey60', fill = 'grey80') +
  geom_jitter(color = 'blue', alpha = .4) +
  geom_abline(intercept = trend_regr_91$coefficients[['(Intercept)']],
              slope = trend_regr_91$coefficients[['pop_trend']],
              color = 'darkred') +
  scale_x_continuous(breaks = c(-1, 0, 1), labels = c('decreasing', 'stable', 'increasing')) +
  coord_cartesian(ylim = c(-.07, .07)) +
  labs(x = 'text population trend',
       y = 'calculated trend from category time series',
       title = 'Pop trend vs change in assessed risk since 1991') +
  annotate('text', x = 0.1, y = -.06, 
           label = paste0('R^2: ',          round(trend_coeffs_91$r.squared, 5), 
                          '\nslope = ',     round(trend_coeffs_91$coefficients[2, 1], 5),
                          '\np (slope) = ', signif(trend_coeffs_91$coefficients[2, 4], 5)),
           hjust = 0)

print(trend_plot_91)

trend_01 <- read_csv(file.path(dir_setup, 'int/trend_calcs/trend_lm_vs_pop_01.csv')) %>%
  filter(!is.na(pop_trend_desc)) %>%
  mutate(pop_trend = case_when(pop_trend_desc == 'decreasing' ~ -1,
                               pop_trend_desc == 'increasing' ~  1,
                               pop_trend_desc == 'stable'     ~  0,
                               TRUE                           ~ NA_real_))

trend_regr_01 <- lm(calc_trend ~ pop_trend, data = trend_01)

trend_coeffs_01 <- summary(trend_regr_01)

### BASED ON FILTERING SINCE 2001 (v3.1)
# Call:
# lm(formula = calc_trend ~ pop_trend, data = trend_01)
# 
# Residuals:
#      Min       1Q   Median       3Q      Max 
# -0.59906 -0.00437  0.00094  0.00455  0.19563 
# 
# Coefficients:
#               Estimate Std. Error t value Pr(>|t|)    
# (Intercept) -0.0009434  0.0006745  -1.399    0.162    
# pop_trend   -0.0053093  0.0008151  -6.514 9.87e-11 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.02199 on 1543 degrees of freedom
#   (969 observations deleted due to missingness)
# Multiple R-squared:  0.02676,	Adjusted R-squared:  0.02613 
# F-statistic: 42.43 on 1 and 1543 DF,  p-value: 9.872e-11


trend_plot_01 <- ggplot(trend_01, aes(x = pop_trend, y = calc_trend, group = pop_trend)) +
  geom_violin(color = 'grey60', fill = 'grey80') +
  geom_jitter(color = 'blue', alpha = .4) +
  geom_abline(intercept = trend_regr_01$coefficients[['(Intercept)']],
              slope = trend_regr_01$coefficients[['pop_trend']],
              color = 'darkred') +
  scale_x_continuous(breaks = c(-1, 0, 1), labels = c('decreasing', 'stable', 'increasing')) +
  coord_cartesian(ylim = c(-.07, .07)) +
  labs(x = 'text population trend',
       y = 'calculated trend from category time series',
       title = 'Pop trend vs change in assessed risk since 2001') +
  annotate('text', x = 0.1, y = -.06, 
           label = paste0('R^2: ',          round(trend_coeffs_01$r.squared, 5), 
                          '\nslope = ',     round(trend_coeffs_01$coefficients[2, 1], 5),
                          '\np (slope) = ', signif(trend_coeffs_01$coefficients[2, 4], 5)),
           hjust = 0)

print(trend_plot_01)

```

## Save species to trend lookup 

The saved trends are based on linear regression of species with at least two non-DD assessments since 1991.  Because of this limitation, only a small number of regionally assessed species contribute to the trend scores.  NOTE:  For OHI, the determination should be made about how to calculate trends - how many assessments must be included? what starting year?

In mapping trends, the mean trend can perhaps be converted more meaningfully into "change in risk category per year" and/or "years per change in risk category".

``` {r save_spp_trends}

### manual values from linear regression:
slope <- trend_coeffs_91$coefficients[2, 1] %>% round(5)
trend_score_lookup <- c('decreasing' = -slope, 'stable' = 0, 'increasing' = slope)

trend_91 <- read_csv(file.path(dir_setup, 'int/trend_calcs/trend_lm_vs_pop_91.csv')) %>%
  dplyr::select(iucn_sid, iucn_rgn, calc_trend) %>%
  distinct()

trend_df <- read_csv(trends_file) %>%
  left_join(spp_risk_ts, by = c('iucn_sid', 'iucn_rgn', 'iucn_version')) %>%
  distinct() %>%
  left_join(trend_91, by = c('iucn_sid', 'iucn_rgn')) %>%
  mutate(trend_score  = ifelse(is.na(calc_trend), trend_score_lookup[pop_trend], calc_trend),
         trend_source = case_when(!is.na(calc_trend)  ~ 'lm', 
                                  !is.na(trend_score) ~ 'regr',
                                  TRUE ~ NA_character_)) %>%
  dplyr::select(iucn_sid, iucn_rgn, pop_trend, trend_score, trend_source) %>%
  distinct()

write_csv(trend_df, file.path(dir_data, sprintf('iucn_trend_by_spp_%s.csv', api_version)))

DT::datatable(trend_df)
```

-----

``` {r prov_footer, results = 'asis'}
prov_wrapup(commit_outputs = FALSE)
```
